# Conflicts:conflic merge

#	.vs/Emision/v16/.suo
#	Contract/obj/Debug/Contract.csproj.CoreCompileInputs.cache
#	Model/obj/Debug/Model.csproj.CoreCompileInputs.cache
#	ServicioEmision/obj/Debug/RdlCompile.cache
#	TestForm/obj/Debug/DesignTimeResolveAssemblyReferences.cache

+++ Model/Sys_Company.cs
@@ -14,7 +14,7 @@ namespace Model
     
     public partial class Sys_Company
     {
-        public int Sys_Id { get; set; }
+        public long Sys_Id { get; set; }
         public string Sys_Rfc { get; set; }
         public string Sys_Name { get; set; }
         public string Sys_FiscalRegime { get; set; }
